chapter16/requirements.txt
# Build/run dependencies
azureml-core~=1.34.0
azureml-sdk~=1.34.0
pandas~=1.3.3
tensorflow~=2.6.0

# Test dependencies
pytest~=7.1.1
pytest-cov~=3.0.0
mock~=4.0.3
tox~=3.24.5